requirements.txt
asgiref==3.8.1
attrs==24.2.0
blinker==1.9.0
build==1.2.2.post1
click==8.1.7
dnspython==2.7.0
email-validator==1.1.3
Flask==2.0.2
Flask-Admin==1.6.0
Flask-Caching==1.10.1
Flask-Cors==5.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
py==1.11.0
pyproject_hooks==1.2.0
pytest==7.0.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
tomli==2.1.0
typing_extensions==4.12.2
Werkzeug==2.0.0
WTForms==3.2.1
